jose and joao feedback
